ECE56800/project/backend/requirements.txt
pyserial==3.2.1
flask==0.12.4
flask-restful==0.2.12
flask-cors==3.0.7
requests==2.18.4
redis==3.2.1
paho-mqtt==1.4.0